FPLA-1754: Added judge to upload SDO journey
