+++ ccd-definition/CaseEventToComplexTypes/draft-SDO/noticeOfProceedings.json
@@ -1,7 +1,7 @@
 [
   {
     "LiveFrom": "01/01/2017",
-    "ID": "NoticeOfProceedings",
+    "ID": "NoticeOfProceedingsSDO",
     "CaseEventID": "draftSDO",
     "CaseFieldID": "noticeOfProceedings",
     "ListElementCode": "proceedingTypes",
