+++ ccd-definition/CaseEventToFields/CareSupervision/draftSDO-nonprod.json
@@ -6,10 +6,21 @@
     "CaseFieldID": "noticeOfProceedings",
     "PageFieldDisplayOrder": 1,
     "DisplayContext": "COMPLEX",
-    "PageID": "noticeOfProceedings",
+    "PageID": "noticeOfProceedingsSDO",
     "CaseEventFieldLabel": " ",
     "PageDisplayOrder": 13,
     "PageColumnNumber": 1,
     "PageShowCondition": "standardDirectionOrder.orderStatus=\"SEALED\""
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "CaseTypeID": "CARE_SUPERVISION_EPO",
+    "CaseEventID": "draftSDO",
+    "CaseFieldID": "judgeAndLegalAdvisor",
+    "PageFieldDisplayOrder": 1,
+    "DisplayContext": "COMPLEX",
+    "PageID": "judgeAndLegalAdvisor",
+    "PageDisplayOrder": 3,
+    "PageColumnNumber": 1
   }
 ]
