+++ ccd-definition/CaseEventToFields/CareSupervision/draftSDO.json
@@ -53,18 +53,6 @@
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "draftSDO",
-    "CaseFieldID": "judgeAndLegalAdvisor",
-    "PageFieldDisplayOrder": 1,
-    "DisplayContext": "COMPLEX",
-    "PageID": "judgeAndLegalAdvisor",
-    "PageDisplayOrder": 3,
-    "PageColumnNumber": 1,
-    "PageShowCondition": "sdoRouter=\"SERVICE\" OR useServiceRoute=\"YES\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "draftSDO",
     "CaseFieldID": "allPartiesHearingDate",
     "PageFieldDisplayOrder": 2,
     "DisplayContext": "READONLY",
