+++ e2e/tests/gatekeeperUploadsSDO_test.js
@@ -15,6 +15,8 @@ BeforeSuite(async (I) => {
 Scenario('Gatekeeper uploads draft standard directions', async (I, caseViewPage, draftStandardDirectionsEventPage) => {
   await caseViewPage.goToNewActions(config.administrationActions.draftStandardDirections);
   await draftStandardDirectionsEventPage.createSDOThroughUpload();
+  await draftStandardDirectionsEventPage.useAllocatedJudge('Bob Ross');
+  await I.retryUntilExists(() => I.click('Continue'), draftStandardDirectionsEventPage.fields.file.preparedSDO);
   await draftStandardDirectionsEventPage.uploadPreparedSDO(config.testPdfFile);
   await draftStandardDirectionsEventPage.markAsDraft();
   await I.completeEvent('Save and continue');
@@ -28,6 +30,8 @@ Scenario('Gatekeeper uploads draft standard directions', async (I, caseViewPage,
 
 Scenario('Gatekeeper uploads final standard directions', async (I, caseViewPage, draftStandardDirectionsEventPage) => {
   await caseViewPage.goToNewActions(config.administrationActions.draftStandardDirections);
+  await draftStandardDirectionsEventPage.useAllocatedJudge('Bob Ross');
+  await I.retryUntilExists(() => I.click('Continue'), draftStandardDirectionsEventPage.fields.file.replacementSDO);
   I.see('mockFile.pdf');
   await draftStandardDirectionsEventPage.uploadReplacementSDO(config.testWordFile);
   await draftStandardDirectionsEventPage.markAsFinal();
